refactor(RM):remove `get_trigger_policies_list` RPC
- Replace the `get_trigger_policies_list` function
implementation using RPC with direct database access
- Remove `get_trigger_policies_list` RPC as it is
not used by any component
