+++ apps/astarte_realm_management/lib/astarte_realm_management/engine.ex
@@ -560,12 +560,6 @@ defmodule Astarte.RealmManagement.Engine do
     end
   end
 
-  def get_trigger_policies_list(realm_name) do
-    _ = Logger.debug("Get trigger policy list", tag: "get_trigger_policies_list")
-
-    Queries.get_trigger_policies_list(realm_name)
-  end
-
   def trigger_policy_source(realm_name, policy_name) do
     _ =
       Logger.debug("Get trigger policy source.",
