+++ apps/astarte_realm_management_api/test/support/helpers/astarte_realm_management_mock.ex
@@ -37,8 +37,6 @@ defmodule Astarte.RealmManagement.API.Helpers.RPCMock do
     Reply,
     UpdateJWTPublicKeyPEM,
     InstallTriggerPolicy,
-    GetTriggerPoliciesList,
-    GetTriggerPoliciesListReply,
     DeleteTriggerPolicy,
     GetTriggerPolicySource,
     GetTriggerPolicySourceReply,
@@ -186,19 +184,6 @@ defmodule Astarte.RealmManagement.API.Helpers.RPCMock do
         generic_error(reason)
         |> ok_wrap
     end
-  end
-
-  defp execute_rpc(
-         {:get_trigger_policies_list,
-          %GetTriggerPoliciesList{
-            realm_name: realm_name
-          }}
-       ) do
-    list = DB.get_trigger_policies_list(realm_name)
-
-    %GetTriggerPoliciesListReply{trigger_policies_names: list}
-    |> encode_reply(:get_trigger_policies_list_reply)
-    |> ok_wrap
   end
 
   defp execute_rpc(
